Decompose monolithic astronomy utils into modular package structure
This change moves apts/utils/astronomy.py into a modular package
apts/utils/astronomy/ to follow the Single Responsibility Principle.
The refraction, alt-az, angular separation, and calculation functions
are split into dedicated submodules and re-exported in the package
__init__.py to maintain complete backward compatibility. All internal
imports have been updated to point directly to the specific submodules,
and a comprehensive unit test suite has been added.
