+++ apts/objects/utils.py
@@ -1,5 +1,5 @@
-from ..utils.astronomy import (
-    calculate_refraction,
+from ..utils.astronomy.refraction import calculate_refraction
+from ..utils.astronomy.calculations import (
     vectorized_geometric_compute,
     vectorized_geometric_imaging_duration,
 )
