+++ apts/utils/astronomy/calculations.py
@@ -5,146 +5,7 @@ import numpy as np
 import pandas as pd
 import pytz
 
-
-def calculate_refraction(alt_deg: Union[float, np.ndarray]) -> Union[float, np.ndarray]:
-    """
-    Calculates atmospheric refraction in degrees using Bennett's formula.
-    Accurate to ~0.02' for altitudes > 0.
-    """
-    alt_deg_arr = np.atleast_1d(alt_deg)
-    refraction_deg = np.zeros_like(alt_deg_arr, dtype=float)
-    # Apply only for objects above -1 degree to avoid tan(0) or instability
-    mask = alt_deg_arr > -1.0
-    if np.any(mask):
-        # R in arcminutes = 1 / tan(h + 7.31 / (h + 4.4))
-        r_arcmin = 1.0 / np.tan(
-            np.deg2rad(alt_deg_arr[mask] + 7.31 / (alt_deg_arr[mask] + 4.4))
-        )
-        refraction_deg[mask] = r_arcmin / 60.0
-
-    return (
-        refraction_deg[0]
-        if np.isscalar(alt_deg) or (isinstance(alt_deg, np.ndarray) and alt_deg.ndim == 0)
-        else refraction_deg
-    )
-
-
-def vectorized_geometric_altaz(
-    lat_deg: float,
-    lon_decimal: float,
-    ras: np.ndarray,
-    decs: np.ndarray,
-    check_times_gmst: Union[float, np.ndarray],
-    sin_dec: Optional[np.ndarray] = None,
-    cd_cr: Optional[np.ndarray] = None,
-    cd_sr: Optional[np.ndarray] = None,
-    return_azimuth: bool = True,
-    return_alt_degrees: bool = True,
-) -> Tuple[np.ndarray, Optional[np.ndarray]]:
-    """
-    Lightning fast geometric altitude and azimuth calculation using vectorized NumPy operations.
-    Supports broadcasting for both single-time and multi-time calculations.
-
-    ras: RA in decimal hours
-    decs: Dec in decimal degrees
-    check_times_gmst: GMST in decimal hours (scalar or 1D array)
-    """
-    # LST calculation
-    lst_hours = check_times_gmst + lon_decimal / 15.0
-    lat_rad = np.deg2rad(lat_deg)
-    lst_rad = np.deg2rad(lst_hours * 15.0)
-
-    sin_lat = np.sin(lat_rad)
-    cos_lat = np.cos(lat_rad)
-
-    # Handling broadcasting for multi-time (check_times_gmst is 1D array)
-    # and multi-object (ras/decs are 1D arrays).
-    if isinstance(lst_rad, np.ndarray) and lst_rad.ndim == 1:
-        # Shapes: Objects (N, 1), Times (1, M)
-        sin_lst = np.sin(lst_rad)[np.newaxis, :]
-        cos_lst = np.cos(lst_rad)[np.newaxis, :]
-        if sin_dec is not None:
-            sin_dec = sin_dec[:, np.newaxis]
-            cd_cr = cd_cr[:, np.newaxis]
-            cd_sr = cd_sr[:, np.newaxis]
-        else:
-            ra_rad = np.deg2rad(ras * 15.0)[:, np.newaxis]
-            dec_rad = np.deg2rad(decs)[:, np.newaxis]
-            sin_dec = np.sin(dec_rad)
-            cd_cr = np.cos(dec_rad) * np.cos(ra_rad)
-            cd_sr = np.cos(dec_rad) * np.sin(ra_rad)
-    else:
-        # Scalar time case
-        sin_lst = np.sin(lst_rad)
-        cos_lst = np.cos(lst_rad)
-        if sin_dec is None:
-            ra_rad = np.deg2rad(ras * 15.0)
-            dec_rad = np.deg2rad(decs)
-            sin_dec = np.sin(dec_rad)
-            cd_cr = np.cos(dec_rad) * np.cos(ra_rad)
-            cd_sr = np.cos(dec_rad) * np.sin(ra_rad)
-
-    # Optimization: cache the intermediate products used by both altitude and azimuth
-    # cd_ch = cos(dec) * cos(hour_angle) = cos(dec) * cos(LST - RA)
-    # Expansion: cos(dec)cos(LST-RA) = cos(LST)cos(dec)cos(RA) + sin(LST)cos(dec)sin(RA)
-    cd_ch = cos_lst * cd_cr + sin_lst * cd_sr
-
-    # sin(alt) = sin(lat)sin(dec) + cos(lat)cos(dec)cos(LST-RA)
-    sin_alt = sin_lat * sin_dec + cos_lat * cd_ch
-
-    if return_alt_degrees:
-        res_alt = np.rad2deg(np.arcsin(np.clip(sin_alt, -1.0, 1.0)))
-    else:
-        # Return raw sin(alt) if degrees are not requested
-        res_alt = sin_alt
-
-    az_deg = None
-    if return_azimuth:
-        # Azimuth calculation using direction cosines
-        # x_cd = cos(H)sin(lat)cos(dec) - sin(dec)cos(lat)
-        x_cd = cd_ch * sin_lat - sin_dec * cos_lat
-        y_cd = sin_lst * cd_cr - cos_lst * cd_sr
-        az_deg = (np.rad2deg(np.arctan2(y_cd, x_cd)) + 180.0) % 360.0
-
-    return res_alt, az_deg
-
-
-def vectorized_angular_separation(
-    ra1: Union[float, np.ndarray],
-    dec1: Union[float, np.ndarray],
-    ra2: Union[float, np.ndarray],
-    dec2: Union[float, np.ndarray],
-    sin_dec1: Optional[np.ndarray] = None,
-    cd_cr1: Optional[np.ndarray] = None,
-    cd_sr1: Optional[np.ndarray] = None,
-) -> Union[float, np.ndarray]:
-    """
-    Calculates angular separation between two sets of coordinates in degrees.
-    """
-    ra1_rad = np.deg2rad(ra1 * 15.0)
-    dec1_rad = np.deg2rad(dec1)
-    ra2_rad = np.deg2rad(ra2 * 15.0)
-    dec2_rad = np.deg2rad(dec2)
-
-    if sin_dec1 is not None:
-        sin_d1 = sin_dec1
-        c_d1_c_r1 = cd_cr1
-        c_d1_s_r1 = cd_sr1
-    else:
-        sin_d1 = np.sin(dec1_rad)
-        c_d1_c_r1 = np.cos(dec1_rad) * np.cos(ra1_rad)
-        c_d1_s_r1 = np.cos(dec1_rad) * np.sin(ra1_rad)
-
-    sin_d2 = np.sin(dec2_rad)
-    cos_d2 = np.cos(dec2_rad)
-    sin_r2 = np.sin(ra2_rad)
-    cos_r2 = np.cos(ra2_rad)
-
-    # cos(sep) = sin(dec1)sin(dec2) + cos(dec1)cos(dec2)cos(ra1-ra2)
-    # Expansion: cos(dec1)cos(dec2)cos(ra1-ra2) = cos(dec2)cos(ra2)cos(dec1)cos(ra1) + cos(dec2)sin(ra2)cos(dec1)sin(ra1)
-    cos_sep = sin_d1 * sin_d2 + cos_d2 * (cos_r2 * c_d1_c_r1 + sin_r2 * c_d1_s_r1)
-
-    return np.rad2deg(np.arccos(np.clip(cos_sep, -1.0, 1.0)))
+from .refraction import calculate_refraction
 
 
 def vectorized_geometric_compute(
